Fix adjoint interface to work with tf custom_grad
tensorflow custom_gradient decorator requires the
grad function to return the gradients as a pair.
The first value should contain the gradients of
all the inputs passed to the function (in our
case this is x0 and t). The second element must
contain the gradients for the model parameters
which are tf.Variable objects stored in the ODE
object. We don't use these params in the function
interface; instead tf passes all the trainable
parameters related to our method in variables
keyword argument.